Updated models to support lazy fetch serialization error, created new endpoints

+++ src/main/java/com/ally/model/auctionitem/AuctionItems.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Builder
 public class AuctionItems {
   private Set<AuctionItem> data;
-  private int elapseTimeMillis;
+  private long elapseTimeMillis;
   private int itemsPerPage = 10;
   private int page;
   private int totalItems;
